feat: 비즈니스 로직에 사용자 예외 발생 부분 추가

+++ src/main/java/com/example/demo/service/BoardService.java
@@ -1,7 +1,14 @@
 package com.example.demo.service;
 
 import java.util.List;
+import java.util.Objects;
 
+import com.example.demo.domain.Article;
+import com.example.demo.exception.BoardHasArticleException;
+import com.example.demo.exception.BoardNotExistException;
+import com.example.demo.exception.BoardNotFoundException;
+import com.example.demo.exception.RequestNullExistException;
+import com.example.demo.repository.ArticleRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,9 +22,11 @@ import com.example.demo.repository.BoardRepository;
 @Transactional(readOnly = true)
 public class BoardService {
 
+    private final ArticleRepository articleRepository;
     private final BoardRepository boardRepository;
 
-    public BoardService(BoardRepository boardRepository) {
+    public BoardService(ArticleRepository articleRepository, BoardRepository boardRepository) {
+        this.articleRepository = articleRepository;
         this.boardRepository = boardRepository;
     }
 
@@ -29,24 +38,46 @@ public class BoardService {
 
     public BoardResponse getBoardById(Long id) {
         Board board = boardRepository.findById(id);
+
+        if(board == null) {
+            throw new BoardNotFoundException();
+        }
+
         return BoardResponse.from(board);
     }
 
     @Transactional
     public BoardResponse createBoard(BoardCreateRequest request) {
         Board board = new Board(request.name());
+
+        if(board.getName() == null) {
+            throw new RequestNullExistException();
+        }
+
         Board saved = boardRepository.insert(board);
         return BoardResponse.from(saved);
     }
 
     @Transactional
     public void deleteBoard(Long id) {
+        List<Article> allArticles = articleRepository.findAll();
+        boolean boardHasArticles = allArticles.stream()
+                .anyMatch(article -> Objects.equals(article.getAuthorId(), id));
+
+        if (boardHasArticles) {
+            throw new BoardHasArticleException();
+        }
         boardRepository.deleteById(id);
     }
 
     @Transactional
     public BoardResponse update(Long id, BoardUpdateRequest request) {
         Board board = boardRepository.findById(id);
+
+        if (board == null) {
+            throw new BoardNotExistException();
+        }
+
         board.update(request.name());
         Board updated = boardRepository.update(board);
         return BoardResponse.from(updated);
