+++ src/main/java/com/example/demo/service/MemberService.java
@@ -1,7 +1,11 @@
 package com.example.demo.service;
 
 import java.util.List;
+import java.util.Objects;
 
+import com.example.demo.domain.Article;
+import com.example.demo.exception.*;
+import com.example.demo.repository.ArticleRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,14 +19,20 @@ import com.example.demo.repository.MemberRepository;
 @Transactional(readOnly = true)
 public class MemberService {
 
+    private final ArticleRepository articleRepository;
     private final MemberRepository memberRepository;
 
-    public MemberService(MemberRepository memberRepository) {
+    public MemberService(ArticleRepository articleRepository, MemberRepository memberRepository) {
+        this.articleRepository = articleRepository;
         this.memberRepository = memberRepository;
     }
 
     public MemberResponse getById(Long id) {
         Member member = memberRepository.findById(id);
+
+        if (member == null) {
+            throw new MemberNotFoundException();
+        }
         return MemberResponse.from(member);
     }
 
@@ -38,17 +48,40 @@ public class MemberService {
         Member member = memberRepository.insert(
             new Member(request.name(), request.email(), request.password())
         );
+
+        if (member.getName() == null || member.getEmail() == null || member.getPassword() == null) {
+            throw new RequestNullExistException();
+        }
         return MemberResponse.from(member);
     }
 
     @Transactional
     public void delete(Long id) {
+        List<Article> allArticles = articleRepository.findAll();
+        boolean memberHasArticles = allArticles.stream()
+                .anyMatch(article -> Objects.equals(article.getAuthorId(), id));
+
+        if (memberHasArticles) {
+            throw new MemberHasArticleException();
+        }
         memberRepository.deleteById(id);
     }
 
     @Transactional
     public MemberResponse update(Long id, MemberUpdateRequest request) {
         Member member = memberRepository.findById(id);
+
+        if (member == null) {
+            throw new MemberNotExistException();
+        }
+
+        boolean emailExists = memberRepository.findAll().stream()
+                .anyMatch(m -> member.getEmail().equals(m.getEmail()));
+
+        if (emailExists) {
+            throw new AlreadyHasEmailException();
+        }
+
         member.update(request.name(), request.email());
         memberRepository.update(member);
         return MemberResponse.from(member);
